cortex/cortex-core/requirements.txt
fastapi>=0.115.11
litellm>=1.63.11
mcp>=1.4.1
passlib[bcrypt]>=1.7.4
pydantic>=2.10.6
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
sqlalchemy>=2.0.39
sqlalchemy-utils>=0.41.1
asyncpg>=0.29.0
sse-starlette>=2.2.1
uvicorn>=0.34.0

# Development dependencies
alembic>=1.15.1
pytest>=8.1.1
httpx>=0.27.0